Seperate effects that work on single environment
Now all things that work on a single environment are moved to the folder
Environment, while things for managing the whole list of environments
stays above in the folder Manager
