+++ src/Manager/Environment/OpenIde.jsx
@@ -35,7 +35,7 @@
  */
 
 import React from 'react';
-import { openSegger } from '../segger';
+import { openSegger } from './segger';
 import Button from './Button';
 
 import environmentPropType from './environmentPropType';
