+++ src/Manager/Environment/segger.js
